refactor: 테이블, 단체 지정 객체 패키지 분리
