+++ src/test/java/kitchenpos/support/fixture/domain/OrderTableFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.support.fixture.domain;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public enum OrderTableFixture {
