+++ src/test/java/kitchenpos/support/fixture/domain/TableGroupFixture.java
@@ -2,8 +2,8 @@ package kitchenpos.support.fixture.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 
 public enum TableGroupFixture {
